+++ model/algorithms/CircleMidPoint.java
@@ -16,13 +16,13 @@ public class CircleMidPoint extends CircleAlgorithm
     }
     
     public void simetria (int xc, int yc, int x, int y, ArrayList<Punto> points) {
-        points.add(new Punto(xc-x, yc+y));//3
-        points.add(new Punto(xc+x, yc-y));//7
-        points.add(new Punto(xc-x, yc-y));//5
-        points.add(new Punto(xc+y, yc+x));//1
-        points.add(new Punto(xc-y, yc+x));//4
-        points.add(new Punto(xc+y, yc-x));//8
-        points.add(new Punto(xc-y, yc-x));//6
+        points.add(new Punto((double)(xc-x), (double)(yc+y)));//3
+        points.add(new Punto((double)(xc+x), (double)(yc-y)));//7
+        points.add(new Punto((double)(xc-x), (double)(yc-y)));//5
+        points.add(new Punto((double)(xc+y), (double)(yc+x)));//1
+        points.add(new Punto((double)(xc-y), (double)(yc+x)));//4
+        points.add(new Punto((double)(xc+y), (double)(yc-x)));//8
+        points.add(new Punto((double)(xc-y), (double)(yc-x)));//6
     }
     
     public ArrayList<Punto> generatePoints (int xC, int yC, int radio) {
@@ -31,7 +31,7 @@ public class CircleMidPoint extends CircleAlgorithm
         x = 0;
         y = radio;
         p = 1 - radio;
-        points.add(new Punto(x+xC, y+yC));
+        points.add(new Punto((double)x+xC, (double)y+yC));
         pValues.add(p);
         simetria(xC, yC, x, y, points);
         while (x < y) {
@@ -42,7 +42,7 @@ public class CircleMidPoint extends CircleAlgorithm
                 y = y - 1;
                 p = p + 2 * (x - y) + 1;
             }
-            points.add(new Punto(x+xC, y+yC));
+            points.add(new Punto((double)x+xC, (double)y+yC));
             pValues.add(p);
             simetria(xC, yC, x, y, points);
         }
