+++ view/Pixel.java
@@ -8,11 +8,11 @@ import java.awt.Cursor;
 
 public class Pixel {
     private int width;
-    private int x, y;
+    private double x, y;
     private Color color;
     private boolean isSelected;
     private view.shapes.ShapeView parentShape;
-    public Pixel(int x, int y, int width, Color color){
+    public Pixel(double x, double y, int width, Color color){
         this.width = width;
         this.x = x;
         this.y = y;
@@ -30,11 +30,11 @@ public class Pixel {
         this.y = y;
     }
     
-    public int getX() {
+    public double getX() {
         return x;
     }
     
-    public int getY() {
+    public double getY() {
         return y;
     }
     
@@ -46,11 +46,11 @@ public class Pixel {
         return parentShape;
     }
     
-    public int getXPlane () {
+    public double getXPlane () {
         return x;
     }
     
-    public int getYPlane () {
+    public double getYPlane () {
         return y;
     }
     
@@ -60,7 +60,7 @@ public class Pixel {
     
     public void paintFill(Graphics g) {
         g.setColor(color);
-        g.fillRect(x*width, y*width, width, width);
+        g.fillRect((int)(x*width), (int)(y*width), width, width);
     }
     
     public void select () {
@@ -75,11 +75,11 @@ public class Pixel {
         if (isSelected) {
             Color selectedColor = new Color(color.getRed(), color.getGreen(), color.getBlue(), 128);
             g.setColor(selectedColor);
-            g.fillRect(x*width-1, y*width-1, width+3, width+3);
+            g.fillRect((int)(x*width-1), (int)(y*width-1), width+3, width+3);
         }
 
         g.setColor(color);
-        g.fillRect(x*width, y*width, width, width);
+        g.fillRect((int)(x*width), (int)(y*width), width, width);
     }
     
     public boolean contiene(int x, int y) {
